requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
avinit==1.3.0
billiard==3.6.4.0
black==22.10.0
boto3==1.26.93
botocore==1.29.93
CacheControl==0.12.11
cachetools==5.2.0
cairocffi==1.3.0
CairoSVG==2.5.2
celery==5.2.7
certifi==2022.9.14
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect2==0.6.0
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.1
django-admin-list-filter-dropdown==1.0.3
django-channels==0.7.0
django-cors-headers==3.11.0
django-encrypted-model-fields==0.6.5
django-filter==21.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-rest-swagger==2.2.0
django-ses==3.3.0
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
dwollav2==2.2.1
fcm-django==1.0.12
firebase-admin==5.4.0
future==0.18.2
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
nulltype==2.3.1
numpy==1.23.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.46
Pillow==9.2.0
plaid-python==9.3.0
platformdirs==2.5.2
plivo==4.33.0
prompt-toolkit==3.0.29
proto-plus==1.22.1
protobuf==4.21.6
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-memcached==1.59
pytz==2022.2.1
redis==4.3.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
twilio==6.63.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
yapf==0.32.0
zope.interface==5.4.0
django-storages==1.13.2
django-ses==3.3.0
